==> 04-extracting-causes-CID10.ipynb <==
# Causas externas de morbidade e de mortalidade: Acidentes de Transporte


A CID-10 é a Classificação Internacional de Doenças [(International Statistical Classification of Diseases and Related Health Problems – ICD](https://icd.who.int/browse10/2010/en), está em sua décima revisão, feita em 2010. 
Essa classificação também abrange as causas externas de mortalidade, como acidentes de trânsito. E visto que o CID-10 também é usado nos dados do SIM, precisamos saber a descrição de cada código a fim de termos maior entendimento sobre a mortalidade no trânsito.

A partir da página [CID-10 Capítulo XX: Causas externas de morbidade e de mortalidade](https://pt.wikipedia.org/wiki/CID-10_Cap%C3%ADtulo_XX:_Causas_externas_de_morbidade_e_de_mortalidade) disponível no Wikipédia, fizemos a extração dos códigos e descrições.

from bs4 import BeautifulSoup as bs
import pandas as pd
import requests
import lxml
pd.set_option('display.max_colwidth', None)

Os dados disponíveis em [A Classificação Internacional de Doenças CID-10 do Mobilidade Ativa]("http://mobilidadeativa.org.br/cid-10/") fornece a descrição das faixas de código. O ideal seria ter a informação no formato mais granular possível.

causes = pd.read_html("http://mobilidadeativa.org.br/cid-10/")
causes[0]

## Extraindo códigos e descrições de forma granular

Vamos tentar a partir da página [CID-10 Capítulo XX: Causas externas de morbidade e de mortalidade do Wikipédia](https://pt.wikipedia.org/wiki/CID-10_Cap%C3%ADtulo_XX:_Causas_externas_de_morbidade_e_de_mortalidade)

page = requests.get("https://pt.wikipedia.org/wiki/CID-10_Cap%C3%ADtulo_XX:_Causas_externas_de_morbidade_e_de_mortalidade")
page = bs(page.text, 'html.parser')

causes = []

for ul in page.findAll('ul'):
    for li in ul.findAll('li'):
        if li.text[0:3] in ['(V0','(V1', '(V2', '(V3', '(V4', '(V5', '(V6', '(V7', '(V8', '(V9']:
            causes.append(li.text.split('\n')[0])

len(causes)

causes_df = pd.DataFrame({'causes': causes})
causes_df.head(15)

causes_df.columns

causes_df = causes_df.drop_duplicates()
len(causes_df)

len(new_causes_df)

pat = r'(\(V\d{2}(?:\.\d)?\)\ )(.*)'
new_causes_df = causes_df['causes'].str.extract(pat=pat).rename(columns={0: 'cod', 1: 'desc'})
new_causes_df.head()

new_causes_df['cod'] = new_causes_df.cod.str.strip()\
.str.replace(pat=r'\(|\)|\.', repl='')\
.str.pad(width=4, side='right', fillchar='0')
new_causes_df.head()

***

## Corrigindo os códigos V19.

Observando a subdivisão do código V19 percebemos que o código `V19.0` se repete mesmo com descrições diferentes. 

new_causes_df.loc[new_causes_df['cod'] == 'V190'].sort_values('desc', ascending=True)

Ao checar na [fonte indicada pela própria página](https://icd.who.int/browse10/2016/en#/V19.0) vimos que os códigos corretos vão de `V19.0` a `V19.9`. 
Fizemos a correção no dataframe e solicitaremos a edição na página do Wikipédia.

| Código | Descrição |
|------|-------------|
|V19.0|Condutor traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente não-de-trânsito|
|V19.1|Passageiro traumatizado em colisão com outro veículos e com veículos não especificados, a motor, em um acidente não-de-trânsito|
|V19.2|Ciclista não especificado traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente não-de-trânsito|
|V19.3|Ciclista (qualquer) traumatizado em um acidente não especificado e não-de-trânsito|
|V19.4|Condutor traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito|
|V19.5|Passageiro traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito|
|V19.6|Ciclista não especificado traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito|
|V19.8|Ciclista (qualquer) traumatizado em outros acidentes de transporte especificados|
|V19.9|Ciclista (qualquer) traumatizado em um acidente de trânsito não especificado|

new_codes = [
{'code': 'V191', 'description': 'Passageiro traumatizado em colisão com outro veículos e com veículos não especificados, a motor, em um acidente não-de-trânsito'},
{'code': 'V192', 'description': 'Ciclista não especificado traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente não-de-trânsito'},
{'code': 'V193', 'description': 'Ciclista (qualquer) traumatizado em um acidente não especificado e não-de-trânsito'},
{'code': 'V194', 'description': 'Condutor traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito'},
{'code': 'V195', 'description': 'Passageiro traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito'},
{'code': 'V196', 'description': 'Ciclista não especificado traumatizado em colisão com outros veículos e com veículos não especificados, a motor, em um acidente de trânsito'},
{'code': 'V198', 'description': 'Ciclista (qualquer) traumatizado em outros acidentes de transporte especificados'},
{'code': 'V199', 'description': 'Ciclista (qualquer) traumatizado em um acidente de trânsito não especificado'}]

new_causes_df['cod']

for new_code in new_codes:
    new_causes_df.loc[(new_causes_df['desc'] == new_code.get('description')) &
             (new_causes_df['cod'] == 'V190'), 'cod'] = new_code.get('code')
    
new_causes_df.loc[27:35]

filepath = './cid10_mapper.json'
new_causes_df.set_index('cod').squeeze().to_json(filepath, orient='index')
